Renamed messages2 folder back to messages
This is part of the cleanup after introducing v2
