+++ src/Frontend/src/components/messages/BodyView.vue
