+++ src/Frontend/src/components/messages/DeleteMessageButton.vue
