+++ src/Frontend/src/components/messages/DiffContent.vue
