+++ src/Frontend/src/components/messages/MessageView.vue
@@ -7,18 +7,18 @@ import FlowDiagram from "./FlowDiagram/FlowDiagram.vue";
 import SequenceDiagram from "./SequenceDiagram.vue";
 import routeLinks from "@/router/routeLinks";
 import { useIsMassTransitConnected } from "@/composables/useIsMassTransitConnected";
-import BodyView from "@/components/messages2/BodyView.vue";
-import HeadersView from "@/components/messages2/HeadersView.vue";
-import StackTraceView from "@/components/messages2/StacktraceView.vue";
+import BodyView from "@/components/messages/BodyView.vue";
+import HeadersView from "@/components/messages/HeadersView.vue";
+import StackTraceView from "@/components/messages/StacktraceView.vue";
 import { useMessageStore } from "@/stores/MessageStore";
-import DeleteMessageButton from "@/components/messages2/DeleteMessageButton.vue";
-import RestoreMessageButton from "@/components/messages2/RestoreMessageButton.vue";
-import RetryMessageButton from "@/components/messages2/RetryMessageButton.vue";
-import EditAndRetryButton from "@/components/messages2/EditAndRetryButton.vue";
-import ExportMessageButton from "@/components/messages2/ExportMessageButton.vue";
+import DeleteMessageButton from "@/components/messages/DeleteMessageButton.vue";
+import RestoreMessageButton from "@/components/messages/RestoreMessageButton.vue";
+import RetryMessageButton from "@/components/messages/RetryMessageButton.vue";
+import EditAndRetryButton from "@/components/messages/EditAndRetryButton.vue";
+import ExportMessageButton from "@/components/messages/ExportMessageButton.vue";
 import TabsLayout from "@/components/TabsLayout.vue";
 import { storeToRefs } from "pinia";
-import MetadataLabel from "@/components/messages2/MetadataLabel.vue";
+import MetadataLabel from "@/components/messages/MetadataLabel.vue";
 import { hexToCSSFilter } from "hex-to-css-filter";
 import SagaDiagram from "./SagaDiagram.vue";
