+++ src/Frontend/src/components/messages/SagaDiagram.vue
