+++ src/Frontend/src/components/messages/SagaDiagram/MessageDataBox.vue
