+++ src/Frontend/src/components/messages/SagaDiagram/NoSagaData.vue
