+++ src/Frontend/src/components/messages/SagaDiagram/SagaCompletedNode.vue
