+++ src/Frontend/src/components/messages/SagaDiagram/SagaHeader.vue
