+++ src/Frontend/src/components/messages/SagaDiagram/SagaOutgoingMessage.vue
