+++ src/Frontend/src/components/messages/SequenceDiagram.vue
