+++ src/Frontend/src/components/messages/SequenceDiagram/EndpointTooltip.vue
