+++ src/Frontend/src/components/messages/SequenceDiagram/EndpointsComponent.vue
