+++ src/Frontend/src/components/messages/SequenceDiagram/HandlersComponent.vue
