+++ src/Frontend/src/components/messages/SequenceDiagram/TimelineComponent.vue
