+++ src/Frontend/src/components/messages/SequenceDiagram/tooltipOverlay.ts
