+++ src/Frontend/src/components/messages/StacktraceFormatter.vue
